Restructure into vendor/drivers/backends/targets; drop PlatformIO
PlatformIO is replaced by CMake throughout. The Teensy firmware is now its own
CMake project under targets/teensy41/, cross-compiled with its own toolchain
file and sharing lib/StaTeX as sources. A cross build cannot be a subdirectory
of the host build because one CMake configure resolves one toolchain. Flags
were taken from `pio run -v` rather than guessed, and the resulting image is
byte-identical to the PlatformIO one.

The tree now expresses one-way dependencies:

  lib/StaTeX/   portable core; declares Graphics2D, knows nothing below it
  drivers/      hardware we drive; knows nothing about the application
  backends/     Graphics2D implementations; the only layer that knows both
  targets/      pins, clocks, linker script, main()
  vendor/       third-party SDKs, verbatim, never edited

TFT_eSPI is gone, replaced by an ILI9488 driver in drivers/ that reaches its
bus through an abstract drivers::DisplayBus which the target implements. It
streams a whole rectangle per address window rather than re-sending
CASET/PASET/RAMWR per pixel, and clips to the panel. The panel is 18-bit only
over SPI, which the driver documents at the point it would otherwise be got
wrong.

Layering fixes:
  - statex_tft.h was a TFT_eSPI backend inside the portable core behind
    #if defined(ARDUINO). Removed.
  - ili9488_graphics.h depended upward on lib/StaTeX and declared its class
    inside namespace statex. Moved to backends/ as statex::backends.
  - statex_record.h was test scaffolding in the shipped core. Moved to
    tests/support/.

Also in here:
  - The SDF sampler's DDA no longer needs 64-bit division, so the only
    external symbol a cortex-m33 build of lib/StaTeX needs is memset. That
    verifies STX-BLD-02, which previously had no test.
  - The STM32U5G9J CubeMX bringup is split into vendor/stm32u5xx and
    targets/stm32u5g9j-dk. CubeMX regenerates its own copy of the ST sources,
    so tools/repoint_cubemx.py re-applies the split; it is idempotent.
  - Dead PlatformIO leftovers removed: include/, firmware.map, the unused
    STATEX_HOST_TEST define, and the OpenFontRender helper (archived beside
    the MicroTeX example it belonged to).

docs/StaTeX-layering.md records why there is deliberately no HAL layer: a
portable Spi/Gpio abstraction collapses as soon as two boards differ in kind
rather than degree, so the portable seam is Graphics2D, defined by what the
application needs rather than by what some hardware happens to be.

+++ lib/StaTeX/statex_sdf.cpp
@@ -70,10 +70,13 @@ bool renderGlyphCoverage(const GlyphRecord& g, const u8* sdfBase, int spread256,
   // --- Texel-space DDA ----------------------------------------------------
   //
   // `u` advances by a constant step per output pixel, so the per-pixel
-  // `(px + 0.5f) / w` becomes an accumulator. Q16 throughout: the step is
-  // computed once in 64-bit, the walk is 32-bit.
-  const i32 du = static_cast<i32>((static_cast<std::int64_t>(sw) << 16) / w);
-  const i32 dv = static_cast<i32>((static_cast<std::int64_t>(sh) << 16) / h);
+  // `(px + 0.5f) / w` becomes an accumulator. Q16 throughout, and 32-bit: the
+  // SDF dimensions are u8, so `255 << 16` is 16.7M and cannot overflow. Using
+  // 64-bit here would be free on the host and two `__aeabi_uldivmod` calls per
+  // glyph on the target, which is the only reason the library would need a
+  // 64-bit divide helper at all.
+  const i32 du = (static_cast<i32>(sw) << 16) / w;
+  const i32 dv = (static_cast<i32>(sh) << 16) / h;
   const i32 u0 = (du >> 1) - 32768;  // pixel centre 0.5 mapped in, minus 0.5
   const i32 v0 = (dv >> 1) - 32768;
